Asset-Management-App/requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
prophet
requests
beautifulsoup4
openpyxl
requests
chardet
Pillow

